+++ MyProject/reward-acceptance-tests/src/test/java/test/classesWithRestTestsMethod/rewardCalculationApp/TariffClassWithMethodsForAcceptanceTests.java
@@ -4,6 +4,8 @@ import io.restassured.response.Response;
 import test.DTOS.TariffDTO;
 import test.rewardapp.requests.CommonRequestForTariffParameters;
 
+import java.math.BigDecimal;
+
 import static io.restassured.RestAssured.given;
 import static org.hamcrest.Matchers.equalTo;
 
@@ -11,7 +13,7 @@ public class TariffClassWithMethodsForAcceptanceTests {
 
     private static final String BASE_URL_REWARD_CALCULATION_APP = "http://localhost:8080";
 
-    public static Response createTariff(String jobType, double amount) {
+    public static Response createTariff(String jobType, BigDecimal amount) {
         CommonRequestForTariffParameters request = CommonRequestForTariffParameters.builder()
                 .tariffDTO(TariffDTO.builder()
                         .jobType(jobType)
