+++ MyProject/reward-payment-app/src/main/java/RewardPayment/JPA/repositories/PaymentRepository.java
@@ -2,8 +2,10 @@ package RewardPayment.JPA.repositories;
 
 import RewardPayment.JPA.domain.Payment;
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.math.BigDecimal;
 import java.util.List;
 public interface PaymentRepository  extends JpaRepository<Payment, Long> {
 
-    List<Payment> findByEmployeeIdAndAmount(Long employeeId, Double amount);
+    List<Payment> findByEmployeeIdAndAmount(Long employeeId, BigDecimal amount);
 }
